+++ src/main/java/org/websoso/WSSServer/notification/dto/FCMTokenRequest.java
@@ -1,4 +1,4 @@
-package org.websoso.WSSServer.dto.user;
+package org.websoso.WSSServer.notification.dto;
 
 import jakarta.validation.constraints.NotBlank;
