+++ src/main/java/org/websoso/WSSServer/notification/dto/PushSettingGetResponse.java
@@ -1,4 +1,4 @@
-package org.websoso.WSSServer.dto.notification;
+package org.websoso.WSSServer.notification.dto;
 
 public record PushSettingGetResponse(
         Boolean isPushEnabled
