+++ src/main/java/org/websoso/WSSServer/notification/dto/PushSettingRequest.java
@@ -1,4 +1,4 @@
-package org.websoso.WSSServer.dto.notification;
+package org.websoso.WSSServer.notification.dto;
 
 import jakarta.validation.constraints.NotNull;
