+++ src/main/java/org/websoso/WSSServer/notification/service/NotificationService.java
@@ -22,11 +22,11 @@ import org.websoso.WSSServer.notification.domain.ReadNotification;
 import org.websoso.WSSServer.user.domain.User;
 import org.websoso.WSSServer.notification.domain.UserDevice;
 import org.websoso.WSSServer.domain.common.NotificationTypeGroup;
-import org.websoso.WSSServer.dto.notification.NotificationCreateRequest;
-import org.websoso.WSSServer.dto.notification.NotificationGetResponse;
-import org.websoso.WSSServer.dto.notification.NotificationInfo;
-import org.websoso.WSSServer.dto.notification.NotificationsGetResponse;
-import org.websoso.WSSServer.dto.notification.NotificationsReadStatusGetResponse;
+import org.websoso.WSSServer.notification.dto.NotificationCreateRequest;
+import org.websoso.WSSServer.notification.dto.NotificationGetResponse;
+import org.websoso.WSSServer.notification.dto.NotificationInfo;
+import org.websoso.WSSServer.notification.dto.NotificationsGetResponse;
+import org.websoso.WSSServer.notification.dto.NotificationsReadStatusGetResponse;
 import org.websoso.WSSServer.exception.exception.CustomNotificationException;
 import org.websoso.WSSServer.exception.exception.CustomNotificationTypeException;
 import org.websoso.WSSServer.notification.FCMClient;
