+++ src/main/java/org/websoso/WSSServer/user/controller/UserController.java
@@ -25,11 +25,10 @@ import org.websoso.WSSServer.application.AuthApplication;
 import org.websoso.WSSServer.user.domain.User;
 import org.websoso.WSSServer.domain.common.SortCriteria;
 import org.websoso.WSSServer.dto.feed.UserFeedsGetResponse;
-import org.websoso.WSSServer.dto.notification.PushSettingGetResponse;
-import org.websoso.WSSServer.dto.notification.PushSettingRequest;
+import org.websoso.WSSServer.notification.dto.PushSettingGetResponse;
+import org.websoso.WSSServer.notification.dto.PushSettingRequest;
 import org.websoso.WSSServer.dto.user.EditMyInfoRequest;
 import org.websoso.WSSServer.dto.user.EditProfileStatusRequest;
-import org.websoso.WSSServer.dto.user.FCMTokenRequest;
 import org.websoso.WSSServer.dto.user.LoginResponse;
 import org.websoso.WSSServer.dto.user.MyProfileResponse;
 import org.websoso.WSSServer.dto.user.NicknameValidation;
